+++ src/LDL/Framework/Base/Collection/Contracts/PrioritySortingInterface.php
@@ -14,8 +14,8 @@ interface PrioritySortingInterface extends SortInterface
      *
      * @param string $order, the order in which to sort, order must be one of:
      *
-     *  SortInterface::SORT_ASCENDING
-     *  SortInterface::SORT_DESCENDING
+     *  Constants::SORT_ASCENDING
+     *  Constants::SORT_DESCENDING
      *
      * @see SortInterface
      * @see PriorityInterface