+++ src/LDL/Framework/Base/Collection/Traits/ReplaceValueByKeyInterfaceTrait.php
@@ -15,6 +15,7 @@ use LDL\Framework\Base\Collection\Contracts\CollectionInterface;
 use LDL\Framework\Base\Collection\Contracts\LockReplaceInterface;
 use LDL\Framework\Base\Collection\Contracts\OnReplaceNoMatchInterface;
 use LDL\Framework\Base\Collection\Contracts\ReplaceValueByKeyInterface;
+use LDL\Framework\Base\Constants;
 use LDL\Framework\Base\Contracts\LockableObjectInterface;
 use LDL\Framework\Helper\ArrayHelper\ArrayHelper;
 use LDL\Framework\Helper\ClassRequirementHelperTrait;
@@ -38,8 +39,8 @@ trait ReplaceValueByKeyInterfaceTrait
         $key,
         $replacement,
         bool $useKeys=true,
-        string $operator = ComparisonOperatorHelper::OPERATOR_SEQ,
-        string $order = ComparisonOperatorHelper::COMPARE_LTR
+        string $operator = Constants::OPERATOR_SEQ,
+        string $order = Constants::COMPARE_LTR
     ) : int
     {
         $this->requireImplements([