+++ src/LDL/Framework/Base/Collection/Traits/SortableScalarInterfaceTrait.php
@@ -4,7 +4,7 @@ namespace LDL\Framework\Base\Collection\Traits;
 
 use LDL\Framework\Base\Collection\Contracts\CollectionInterface;
 use LDL\Framework\Base\Collection\Contracts\SortableScalarInterface;
-use LDL\Framework\Base\Collection\Contracts\SortInterface;
+use LDL\Framework\Base\Constants;
 use LDL\Framework\Helper\ClassRequirementHelperTrait;
 
 /**
@@ -17,7 +17,7 @@ trait SortableScalarInterfaceTrait
     use ClassRequirementHelperTrait;
 
     //<editor-fold desc="SortableScalarInterface methods">
-    public function sortScalar(string $sort=SortInterface::SORT_ASCENDING): CollectionInterface
+    public function sortScalar(string $sort=Constants::SORT_ASCENDING): CollectionInterface
     {
         $this->requireImplements([
             CollectionInterface::class,
@@ -31,7 +31,7 @@ trait SortableScalarInterfaceTrait
             return is_scalar($v);
         })
         ->sort(static function($a, $b) use($sort){
-            return SortInterface::SORT_ASCENDING === $sort ? $a <=> $b : $b <=> $a;
+            return Constants::SORT_ASCENDING === $sort ? $a <=> $b : $b <=> $a;
         });
     }
     //</editor-fold>