+++ src/LDL/Framework/Helper/IterableHelper.php
@@ -3,24 +3,12 @@
 namespace LDL\Framework\Helper;
 
 use LDL\Framework\Base\Collection\Contracts\ComparisonInterface;
+use LDL\Framework\Base\Constants;
 
 final class IterableHelper
 {
     public const BEFORE_LAST_POSITION = -2;
     public const LAST_POSITION = -1;
-
-    public const PHP_TYPE_STRING = 'string';
-    public const PHP_TYPE_BOOL = 'boolean';
-    public const PHP_TYPE_ARRAY = 'array';
-    public const PHP_TYPE_INTEGER = 'integer';
-    public const PHP_TYPE_DOUBLE = 'double';
-    public const PHP_TYPE_OBJECT = 'object';
-    public const PHP_TYPE_RESOURCE = 'resource';
-    public const PHP_TYPE_NULL = 'null';
-
-    public const LDL_TYPE_STRNUM = 'strnum';
-    public const LDL_TYPE_UINT = 'uint';
-    public const LDL_TYPE_UDOUBLE = 'udouble';
 
     /**
      * @param iterable $items
@@ -142,15 +130,15 @@ final class IterableHelper
 
         foreach($types as $type){
             switch(strtolower($type)){
-                case self::LDL_TYPE_UINT:
+                case Constants::LDL_TYPE_UINT:
                     $hasUint = true;
                     break;
 
-                case self::LDL_TYPE_UDOUBLE:
+                case Constants::LDL_TYPE_UDOUBLE:
                     $hasUDouble = true;
                     break;
 
-                case self::LDL_TYPE_STRNUM:
+                case Constants::LDL_TYPE_STRNUM:
                     $hasStrNum = true;
                     break;
             }
@@ -159,15 +147,15 @@ final class IterableHelper
         return self::filter($items, static function ($val) use ($types, $hasUint, $hasUDouble, $hasStrNum){
             $type = strtolower(gettype($val));
 
-            if($hasUint && self::PHP_TYPE_INTEGER === $type){
+            if($hasUint && Constants::PHP_TYPE_INTEGER === $type){
                 return $val >= 0;
             }
 
-            if($hasUDouble && self::PHP_TYPE_DOUBLE === $type){
+            if($hasUDouble && Constants::PHP_TYPE_DOUBLE === $type){
                 return $val >= 0;
             }
 
-            if($hasStrNum && (self::PHP_TYPE_INTEGER === $type || self::PHP_TYPE_DOUBLE === $type)){
+            if($hasStrNum && (Constants::PHP_TYPE_INTEGER === $type || Constants::PHP_TYPE_DOUBLE === $type)){
                 return true;
             }